#36 Wrote semester earnings calculator
I wrote the semesters earning calculator to take out tax and rent. Placeholder functions were used for the tax percent and semester rent as those functions are not completed to my knowledge. I may need to rework this depending on how the tax grabber and cost of living calculator work when completed.